--- packages/design-system/src/types/Tooltip/Tooltip.d.ts
@@ -1,84 +1,0 @@
-import * as React from 'react';
-
-interface TooltipProps {
-  /**
-   * Classes applied to the tooltip trigger when the tooltip is active
-   */
-  activeClassName?: string;
-  /**
-   * Helpful description of the tooltip for screenreaders
-   */
-  ariaLabel?: string;
-  /**
-   * Tooltip trigger content
-   */
-  children: React.ReactNode;
-  /**
-   * When provided, this will render the passed in component for the tooltip trigger. Typically this will be a `button`, `a`, or rarely an `input` element.
-   */
-  component?: React.ReactElement<any> | any | ((...args: any[]) => any);
-  /**
-   * Classes applied to the tooltip trigger
-   */
-  className?: string;
-  /**
-   * Tooltip that behaves like a dialog, i.e. a tooltip that only appears on click, traps focus, and contains interactive content. For more information, see Deque's [tooltip dialog documentation](https://dequeuniversity.com/library/aria/tooltip-dialog)
-   */
-  dialog?: boolean;
-  /**
-   * `id` applied to tooltip body container element. If not provided, a unique id will be automatically generated and used.
-   */
-  id?: string;
-  /**
-   * Sets the size of the invisible border around interactive tooltips that prevents it from immediately hiding when the cursor leaves the tooltip.
-   */
-  interactiveBorder?: number;
-  inversed?: boolean;
-  /**
-   * Applies `skidding` and `distance` offsets to the tooltip relative to the trigger. See the [`popperjs` docs](https://popper.js.org/docs/v2/modifiers/popper-offsets/) for more info.
-   */
-  offset?: [number, number];
-  /**
-   * Called when the tooltip is hidden
-   */
-  onClose?: () => void;
-  /**
-   * Called when the tooltip is shown
-   */
-  onOpen?: () => void;
-  /**
-   * Placement of the tooltip body relative to the trigger. See the [`popperjs` docs](https://popper.js.org/docs/v2/constructors/#options) for more info.
-   */
-  placement?: 'auto' | 'bottom' | 'top' | 'right' | 'left';
-  /**
-   * `maxWidth` styling applied to the tooltip body
-   */
-  maxWidth?: string;
-  /**
-   * Content inside the tooltip body or popover. If this contains interactive elements use the `dialog` prop.
-   */
-  title: React.ReactNode;
-  /**
-   * Duration of the `react-transition-group` CSSTransition. See the [`timeout` option](http://reactcommunity.org/react-transition-group/transition#Transition-prop-timeout) for more info.
-   */
-  transitionDuration?: number;
-  /**
-   * `zIndex` styling applied to the tooltip body
-   */
-  zIndex?: number;
-}
-
-interface TooltipState {
-  active: boolean,
-  isHover: boolean,
-  isMobile: boolean,
-}
-
-type OmitProps = 'title';
-
-export default class Tooltip extends React.Component<
-  Omit<React.ComponentPropsWithRef<'button'>, OmitProps> & TooltipProps,
-  TooltipState
-> {
-  render(): JSX.Element;
-}
