--- packages/design-system/src/types/Tooltip/TooltipIcon.d.ts
@@ -1,9 +1,0 @@
-import * as React from 'react';
-
-export interface TooltipIconProps {
-  inversed?: boolean;
-}
-
-declare const TooltipIcon: React.FC<TooltipIconProps>;
-
-export default TooltipIcon;
