+++ packages/design-system/src/types/index.d.ts
@@ -6,8 +6,6 @@ export { default as HelpDrawer } from './HelpDrawer/HelpDrawer';
 export { default as HelpDrawerToggle } from './HelpDrawer/HelpDrawerToggle';
 export { default as SkipNav } from './SkipNav/SkipNav';
 export { default as StepList } from './StepList/StepList';
-export { default as Tooltip } from './Tooltip/Tooltip';
-export { default as TooltipIcon } from './Tooltip/TooltipIcon';
 
 // Type definitions generated by cmsds scripts
 export * from './Accordion';
@@ -30,6 +28,7 @@ export * from './Table';
 export * from './Tabs';
 export * from './TextField';
 export * from './Tooltip';
+export * from './TooltipIcon';
 export * from './Icons';
 export * from './UsaBanner';
 export * from './VerticalNav';
