+++ packages/ds-medicare-gov/docs/src/pages/components/Tooltip/_Tooltip.docs.scss
@@ -19,7 +19,7 @@ This component makes use of [`@popperjs/core`](https://popper.js.org/docs/v2/) f
 
 @react-example Tooltip.example.jsx
 
-@react-props Tooltip.jsx
+@react-props Tooltip.tsx
 
 Style guide: components.tooltip.tooltip
